Move out the StripeEvent struct
